chore: initial split of about container
