+++ src/gui/elements/mod.rs
@@ -1,0 +1,6 @@
+mod about;
+pub use about::about_container;
+
+// Default values used on multiple elements.
+pub static DEFAULT_FONT_SIZE: u16 = 14;
+pub static DEFAULT_PADDING: u16 = 10;
